Add CLI
Adds a CLI to set up the render and pick between default scenes.

Eventually, this CLI will likely need to be reworked to take in
for example a JSON file which specifies the scene (I think
the camera related arguments would go in such a file).

But for now, these arguments help development.