Fix issue where scenes could be too bright
Fixes an issue where color space confusion in the color accumulator
across samples caused final renders to be too bright.

We were converting from our linear colors (in a DVec3) to Srgb,
then back to linear for each ray before adding to the color
accumulator. Now, we just stay in linear (DVec3) until the
end, after we divide by the sample count, to convert
to Srgb.

I'm not totally sure why this was resulting in values that were
too bright. Maybe the many conversions were causing error to
build up, or maybe there is a detail re: color spaces I'm
missing here. But, this is more in line with the textbook,
so we will go with this for now.