Change to f32 implementation
The HRPP algorithm uses a hash on f32. We change to be consistent
with that.
Also, most ray tracers use f32 - double is only necessary for
some specific visual artifacts that I don't want to worry about
(for me, I used f64 due to some visual striping on very large
implicit spheres). But those are so rarely rendered that the
performance gains of f32 is worth it.